+++ examples/langchain_examples/01-rag-LangChain-all-sources.py
@@ -12,7 +12,8 @@ from langchain.document_loaders import ConfluenceLoader
 from langchain_aws import BedrockEmbeddings, ChatBedrock
 from langchain_community.vectorstores import FAISS
 from langchain_core.documents import Document
-from langchain_google_community import GoogleDriveLoader  # type: ignore[import-untyped]
+from langchain_google_community import \
+    GoogleDriveLoader  # type: ignore[import-untyped]
 from pangea_multipass import ConfluenceME, GDriveAPI, GDriveME, enrich_metadata
 from pangea_multipass_langchain import (ConfluenceAuth, DocumentFilterMixer,
                                         LangChainConfluenceFilter,