+++ examples/llama_index_examples/05-github-check.py
@@ -3,7 +3,8 @@
 
 import os
 
-from pangea_multipass import GithubReader, PangeaMetadataKeys  # type: ignore[attr-defined]
+from pangea_multipass import GitHubReader  # type: ignore[attr-defined]
+from pangea_multipass import PangeaMetadataKeys
 from pangea_multipass_llama_index import (LlamaIndexGitHubProcessor,
                                           from_multipass)
 
@@ -11,7 +12,7 @@ from pangea_multipass_llama_index import (LlamaIndexGitHubProcessor,
 admin_token = os.getenv("GITHUB_ADMIN_TOKEN")
 assert admin_token
 
-reader = GithubReader(admin_token)
+reader = GitHubReader(admin_token)
 documents = reader.load_data()
 print(f"Loaded {len(documents)} docs:")