+++ examples/multipass_examples/03-slack-check-user-token.py
@@ -2,8 +2,8 @@
 # Author: Pangea Cyber Corporation
 
 import os
-from pangea_multipass import SlackReader
 
+from pangea_multipass import SlackReader
 
 admin_token = os.getenv("SLACK_ADMIN_TOKEN")
 assert admin_token
@@ -15,10 +15,10 @@ print(f"Loaded {len(documents)} messages.")
 # Inference time
 from pangea_multipass import SlackProcessor, get_document_metadata
 
-user_email = os.getenv("SLACK_USER_EMAIL")
-assert user_email
+user_token = os.getenv("SLACK_USER_TOKEN")
+assert user_token
 
-processor = SlackProcessor(token=admin_token, get_node_metadata=get_document_metadata, user_email=user_email)
+processor = SlackProcessor(token=user_token, get_node_metadata=get_document_metadata)
 filter = processor.get_filter()
 print("User has acess to channel ids:")
 for id in filter.value:
@@ -26,4 +26,3 @@ for id in filter.value:
 
 filtered_docs = processor.filter(nodes=documents)
 print(f"User has access to {len(filtered_docs)} messages")
-